Moved configuration file to app folder. Atlas generator now creates raw folder if missing.

+++ src/main/java/de/bushnaq/abdalla/engine/util/AtlasGenerator.java
@@ -63,7 +63,10 @@ public class AtlasGenerator {
         File aggregatedInputFolderFile = new File(aggregatedInputFolder);
         logger.info("----------------------------------------------------------------------------------");
         logger.info(generationReason);
-        FileUtils.cleanDirectory(aggregatedInputFolderFile);
+        if (!aggregatedInputFolderFile.exists())
+            aggregatedInputFolderFile.mkdir();
+        else
+            FileUtils.cleanDirectory(aggregatedInputFolderFile);
         atlasFile.delete();
         (new File(atlasAtlasFileName)).delete();
         generateFonts();
